Kevin Mack "Making a Successful Component Library: Digital Brand Style Guides" slides
Creating a component library is not a new piece of the development and design process but with the growth of responsive and contextual factors it has become a requirement for complex systems. Digital Style Guides bring alignment of teams and solve complexities by providing quality tested code that is fully reusable across all projects. Within this presentation, we will discuss the variations of different style guides, toolkits, and frameworks that exist today and identify the right solution for your team to create a successful component library.